chore: rename tsx files and type files to be pascal case
