+++ ui/components/flows/Chat.tsx
@@ -10,13 +10,13 @@
 import { useChat } from '@ai-sdk/react'
 import { DefaultChatTransport } from 'ai'
 import { useState, useCallback, useMemo, useEffect, useRef } from 'react'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { ScrollArea } from '@/components/ui/scroll-area'
+import { Button } from '@/components/ui/Button'
+import { Input } from '@/components/ui/Input'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/Card'
+import { ScrollArea } from '@/components/ui/ScrollArea'
 import { Send, Bot, Loader2, MessageSquarePlus, Paperclip} from 'lucide-react'
-import { apiClient, type FlowInfo } from '@/lib/api-client'
-import { type FileAttachment } from '@/types/flow'
+import { apiClient, type FlowInfo } from '@/lib/apiClient'
+import { type FileAttachment } from '@/types/Flow'
 import MessageBubble from '@/components/chat/MessageBubble'
 import AttachmentDisplay from '@/components/chat/AttachmentDisplay'
