+++ ui/components/inputs/DatePickerInput.tsx
@@ -5,15 +5,15 @@ import { CalendarIcon } from "lucide-react"
 import { format } from "date-fns"
 
 import { cn } from "@/lib/utils"
-import { Button } from "@/components/ui/button"
-import { Calendar } from "@/components/ui/calendar"
+import { Button } from "@/components/ui/Button"
+import { Calendar } from "@/components/ui/Calendar"
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/components/ui/popover"
+} from "@/components/ui/Popover"
 
-import type { SchemaProperty, FlowInputValue } from '../../types/flow'
+import type { SchemaProperty, FlowInputValue } from '../../types/Flow'
 
 interface DatePickerInputProps {
   name: string
