+++ ui/components/inputs/NumberInput.tsx
@@ -6,8 +6,8 @@
 
 'use client'
 
-import { Input } from '@/components/ui/input'
-import type { SchemaProperty, FlowInputValue } from '../../types/flow'
+import { Input } from '@/components/ui/Input'
+import type { SchemaProperty, FlowInputValue } from '../../types/Flow'
 
 interface NumberInputProps {
   name: string
